Move SndMmap and SndContext into separate files, move openAudioFile into TrackUtils
