+++ noisereduction/NoiseReduction.h
@@ -8,7 +8,6 @@
 #pragma once
 
 #include <memory>
-#include "Utils.h"
 #include "InputTrack.h"
 #include "OutputTrack.h"
