+++ noisereduction/SndContext.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <sndfile.h>
+
+struct SndContext {
+    SNDFILE* file;
+    SF_INFO info;
+};
